+++ ios/CleverTapReact/CleverTapSDK.framework/Headers/CleverTapBuildInfo.h
@@ -1,0 +1,10 @@
+//
+//  CleverTapBuildInfo.h
+//  CleverTapSDK
+//
+//  Copyright (c) 2015 WizRocket. All rights reserved.
+//
+
+#define WR_SDK_VERSION 1
+#define WR_SDK_REVISION 19740
+#define WR_SDK_RELEASE_DATE 20170130