+++ ios/CleverTapReact/CleverTapSDK.framework/Headers/CleverTapEventDetail.h
@@ -1,0 +1,10 @@
+#import <Foundation/Foundation.h>
+
+@interface CleverTapEventDetail : NSObject
+
+@property (nonatomic, strong) NSString *eventName;
+@property (nonatomic) NSTimeInterval firstTime;
+@property (nonatomic) NSTimeInterval lastTime;
+@property (nonatomic) int count;
+
+@end