+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 0.1.2 *(16 February, 2017)*
+-------------------------------------------
+*(Supports CleverTap 3.1.2 and React Native v0.41.2)*
+
+Add CleverTapSDK framework as zip + postinstall unzip, as npm doesn't like iOS framework symlinks.
+
 Version 0.1.1 *(15 February, 2017)*
 -------------------------------------------
 *(Supports CleverTap 3.1.2 and React Native v0.41.2)*